refactor(library): extract track-preference queries into a service
